+++ ic-agent/src/agent/mod.rs
@@ -185,7 +185,7 @@ impl Agent {
     /// Create an instance of an [`Agent`].
     pub fn new(config: agent_config::AgentConfig) -> Result<Agent, AgentError> {
         let client = config.http_service.unwrap_or_else(|| {
-            Arc::new(Retry429Logic {
+            Arc::new(RetryLogic {
                 client: config.client.unwrap_or_else(|| {
                     #[cfg(not(target_family = "wasm"))]
                     {
@@ -2237,13 +2237,13 @@ where
 }
 
 #[derive(Debug)]
-struct Retry429Logic {
+struct RetryLogic {
     client: Client,
 }
 
 #[cfg_attr(target_family = "wasm", async_trait(?Send))]
 #[cfg_attr(not(target_family = "wasm"), async_trait)]
-impl HttpService for Retry429Logic {
+impl HttpService for RetryLogic {
     async fn call<'a>(
         &'a self,
         req: &'a (dyn Fn() -> Result<http::Request<Bytes>, AgentError> + Send + Sync),
@@ -2275,6 +2275,9 @@ impl HttpService for Retry429Logic {
                     crate::util::sleep(Duration::from_millis(250)).await;
                     continue;
                 }
+            } else if resp.status() == StatusCode::SERVICE_UNAVAILABLE {
+                crate::util::sleep(Duration::from_millis(250)).await;
+                continue;
             } else {
                 break Ok(resp);
             }